pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pystackql"
version = "3.8.1"
description = "A Python interface for StackQL"
readme = "README.rst"
authors = [
    {name = "Jeffrey Aven", email = "javen@stackql.io"}
]

license = "MIT"
classifiers = [
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.9"
dependencies = [
    "requests",
    "pandas",
    "IPython",
    "psycopg[binary]>=3.1.0",
    "nest-asyncio>=1.5.5",
    "termcolor>=1.1.0",
    "tqdm>=4.61.0",
    "PyYAML>=5.4.0",
]

[tool.setuptools.packages.find]
include = ["pystackql", "pystackql.*"]

[tool.setuptools]
include-package-data = true

[project.urls]
Homepage = "https://github.com/stackql/pystackql"
Documentation = "https://pystackql.readthedocs.io"